+++ ThirdPartyAdapters/inmobi/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## InMobi Android Mediation Adapter Changelog
 
+#### Next Version
+- Maps `AgeRestrictedTreatment` to InMobi's COPPA API.
+
 #### Version 11.3.0.0
 - Verified compatibility with InMobi Kotlin SDK version 11.3.0.